==> Code/1 - Preview Raw Data.ipynb <==
## Preview Raw Data

### Import Packages

import pandas as pd
import numpy as np
import datetime as dt

### Read in CSVs, Preview, and Check Formatting

#### CRSP Monthly Stock Event - Name History

qcrspmse_raw = pd.read_csv('qcrspmse_raw.csv.gz', compression='gzip')

qcrspmse_raw.head()

qcrspmse_raw.dtypes

#### (Monthly) S&P 500 Index

qcrsp500_raw = pd.read_csv('qcrsp500_raw.csv.gz', compression='gzip')

qcrsp500_raw.head()

qcrsp500_raw.dtypes

#### (Daily) S&P 500 Index

qcrsp500d_raw = pd.read_csv('qcrsp500d_raw.csv.gz', compression='gzip')

qcrsp500d_raw.head()

qcrsp500d_raw.dtypes

#### CRSP Monthly Stock File

qcrspmsf_raw = pd.read_csv('qcrspmsf_raw.csv.gz', compression='gzip')

qcrspmsf_raw.head()

qcrspmsf_raw.dtypes

#### CRSP Monthly Stock Event - Delisting

qdlret_raw = pd.read_csv('qdlret_raw.csv.gz', compression='gzip')

qdlret_raw.head()

qdlret_raw.dtypes

#### CRSP Daily Stock Event - Delisting

qdlretd_raw = pd.read_csv('qdlretd_raw.csv.gz', compression='gzip')

qdlretd_raw.head()

qdlretd_raw.dtypes

### Appendix

#CRSP Daily Stock Event - Delisting
try:
    dlret=pickle.load(open("qdlretd_raw.p","rb"))
except:
    db = wrds.Connection(wrds_username='natejensen')
    dlret = db.raw_sql("""select * from crspq.dsedelist""")
    db.close()
    #pickle.dump(dlret,open("qdlretd_raw.p","wb"))
    
# write a pandas dataframe to gzipped CSV file
dlret.to_csv("qdlretd_raw.csv.gz", 
           index=False, 
           compression="gzip")
