==> Code/2a - Data Pre-Processing & Investible Universe.ipynb <==
## Pre-Processing & Investible Universe

Notes:
Program 2a does a bit of preprocessing/cleaning, but the most important thing is how it limits the investable universe that we examine
* Working back from the end goal, we (arbitrarily) want ~2000 stocks (sorted by market cap) to divide up into quintiles at the end
* So we create a universe that has 3000 stocks every year just in case
  * But because we are looking at previous year’s beta to predict the next year’s returns, we need to make sure that stock exists in both years
  * Stocks will enter and leave the top 3000 for various reasons
  * So if you look at the total number of stocks that come out of permscreen, it is much larger than the 3000 ever year, but every year you are guaranteed to have 3000 stocks to work with that have a beta signal that year, as well as return data the year after
* **Very specific to asam**, since we invest in December, we picked stocks that have returns in end of December, for predicting next year’s returns

### Import Packages

import pandas as pd
import numpy as np
import datetime as dt
import psycopg2 
import matplotlib.pyplot as plt
from dateutil.relativedelta import *
from pandas.tseries.offsets import *
from scipy import stats
import statsmodels.api as sm
import statistics
import sys
sys.path.insert(0, "../")
import util
import multiprocessing as mp
from  timeit import default_timer as timer

### Set Local Macro Variables

#num of top market equity to keep
numstocks=3000

### Read in util.py Function Bank

from IPython.lib.backgroundjobs import BackgroundJobFunc

with open('util.py') as code:
    job = BackgroundJobFunc(exec, code.read())

result = job.run()

### Import Data

#### CRSP Monthly Stock File

crsp_m = pd.read_csv('qcrspmsf_raw.csv.gz', compression='gzip')

crsp_m = crsp_m[['permno', 'permco', 'date', 'ret', 'retx', 'shrout', 'prc']]

crsp_m.head()

#### CRSP Monthly Stock Event - Delisting

qcrspmse_raw = pd.read_csv('qcrspmse_raw.csv.gz', compression='gzip')

qcrspmse_raw = qcrspmse_raw[['permno', 'shrcd', 'exchcd', 'namedt', 'nameendt']]

qcrspmse_raw = qcrspmse_raw[qcrspmse_raw['exchcd'].isin([1, 2, 3])]
qcrspmse_raw = qcrspmse_raw[qcrspmse_raw['shrcd'].isin([10, 11])]

qcrspmse_raw.head()

##### Join monthly stock data with name history

crsp_m = crsp_m.merge(qcrspmse_raw, on='permno', how='left')

crsp_m = crsp_m[(crsp_m.namedt <= crsp_m.date) & (crsp_m.date <= crsp_m.nameendt)]

crsp_m[['permco','permno','shrcd','exchcd']] = crsp_m[['permco','permno','shrcd','exchcd']].astype(int)

crsp_m.head()

#### CRSP Monthly Stock Event - Delisting

qdlret_raw = pd.read_csv('qdlret_raw.csv.gz', compression='gzip')

dlret = qdlret_raw[['permno', 'dlret', 'dlstdt']]

dlret[['permno']] = dlret[['permno']].astype(int)

dlret['date'] = pd.to_datetime(dlret['dlstdt'])

dlret.head()

##### Format Dates

# Set all timestamps to common month end
crsp_m = util.monthEnd(crsp_m)
dlret = util.monthEnd(dlret)

### Combine Returns

crsp = util.comebineRet(crsp_m,dlret)

### Prune Stocks

# Prune stocks, map all equity to largest permno
crsp=util.maxCap(crsp)

# Only keeping top X stocks by market cap per jdate, so a stock could be missing from long time period ago
crsp=util.topX(crsp,'me',numstocks)

crsp.head()

# EXPORT MONTHLY PERMNO LIST FOR DAILY CALC HERE
# will reduce previous list to AT MOST 3k stocks per year effectively,
# based on the 3k that exist in December of that year
# if stocks did not exist jan of that year, then jan should have fewer 
# than 3k stocks.
crsp['month']=crsp['jdate'].dt.month
permlist=crsp[crsp['month']==12][['permno','jdate']].drop_duplicates()

permlist.to_csv("qpermlist.csv.gz", 
           index=False, 
           compression="gzip")

# Dump pruned crps that matches permlist
crsp.to_csv("qlocalcrsp_matched.csv.gz", 
           index=False, 
           compression="gzip")

Notes:
* Filter to only keep stocks with December return
* In order to compare vs prior year
* Delisting events is to check for existence
