==> Code/3a - Win Beta Calculation.ipynb <==
## Win Beta Calculation

Program 31 performs win beta calc. 
It’s very time consuming because of the rolling window + covariance.

### Import Packages

import pandas as pd
import numpy as np
import datetime as dt
import wrds
import psycopg2 
import matplotlib.pyplot as plt
from dateutil.relativedelta import *
from pandas.tseries.offsets import *
from scipy import stats
import statsmodels.api as sm
import statistics
import pickle
import sys
sys.path.insert(0, "../")
import util
import multiprocessing as mp
from  timeit import default_timer as timer
from importlib import reload
reload(util)

### Read in util.py Function Bank

from IPython.lib.backgroundjobs import BackgroundJobFunc

with open('util.py') as code:
    job = BackgroundJobFunc(exec, code.read())

result = job.run()

### Set Local Macro Variables

# Num of top market equity to keep
numstocks=3000

# Trailing window NOW IN YEARS
window=1

### Import Data

#### Daily CRSP File

crsp_d = pd.read_csv('qcrspdsf_raw.csv.gz', compression='gzip', usecols=lambda x: x.lower())

crsp_d.columns = crsp_d.columns.str.lower()
crsp_d = crsp_d[['permno', 'permco', 'date', 'ret', 'retx', 'shrout', 'prc']]
crsp_d[['permno','permco']] = crsp_d[['permno','permco']].astype(int)
crsp_d['ret'] = pd.to_numeric(crsp_d['ret'], errors='coerce')
crsp_d['retx'] = pd.to_numeric(crsp_d['retx'], errors='coerce')

crsp_d['date'] = pd.to_datetime(crsp_d['date'])

crsp_d.head()

crsp_d.dtypes

#### CRSP Daily Stock Event - Delisting

qcrspdse_raw = pd.read_csv('qcrspdse_raw.csv.gz', compression='gzip')

qcrspdse_raw = qcrspdse_raw[['permno', 'shrcd', 'exchcd', 'namedt', 'nameendt']]

qcrspdse_raw[['permno','shrcd','exchcd']] = qcrspdse_raw[['permno','shrcd','exchcd']].astype(int)
qcrspdse_raw = qcrspdse_raw[qcrspdse_raw['exchcd'].isin([1, 2, 3])]
qcrspdse_raw = qcrspdse_raw[qcrspdse_raw['shrcd'].isin([10, 11])]

qcrspdse_raw['namedt'] = pd.to_datetime(qcrspdse_raw['namedt'])
qcrspdse_raw['nameendt'] = pd.to_datetime(qcrspdse_raw['nameendt'])

qcrspdse_raw.head()

##### Join daily stock data with name history

crsp_d = crsp_d.merge(qcrspdse_raw, on='permno', how='left')

crsp_d = crsp_d[(crsp_d.namedt <= crsp_d.date) & (crsp_d.date <= crsp_d.nameendt)]

crsp_d.head()

#### S&P 500 Daily Index

crsp_sp = pd.read_csv('qcrsp500d_raw.csv.gz', compression='gzip')

crsp_sp = crsp_sp.rename(columns={'caldt':'date'})
crsp_sp['date'] = pd.to_datetime(crsp_sp['date'])

crsp_sp.head()

#### CRSP Daily Stock Event - Delisting

dlret = pd.read_csv('qdlretd_raw.csv.gz', compression='gzip')

dlret = dlret[['permno', 'dlret', 'dlstdt']]

dlret[['permno']] = dlret[['permno']].astype(int)

dlret['date'] = pd.to_datetime(dlret['dlstdt'])

dlret.head()

##### Format Dates

crsp_d = util.jdate(crsp_d)
crsp_sp = util.jdate(crsp_sp)
dlret = util.jdate(dlret)

### Trim Stocks

#### Import Saved Permno from Program 2a (aka Investible Universe)

permlist = pd.read_csv('qpermlist.csv.gz', compression='gzip')

permlist['jdate'] = pd.to_datetime(permlist['jdate'])

permlist['month']=permlist['jdate'].dt.month
permlist['year']=permlist['jdate'].dt.year

# Only keep permno that show up in top 3000 in December of that year,
permlist=permlist[permlist['month']==12]

print(permlist)

# Used for trimming crsp_d
permshort=(permlist['permno'].drop_duplicates())

print(permshort)

# Trim to last 30 years
crsp_d['year'] = crsp_d['jdate'].dt.year
crsp_d = crsp_d[crsp_d['year']>=1990]
permlist['year'] = permlist['jdate'].dt.year
permlist = permlist[permlist['year']>=1990]

crsp_d.head()

### Combine Returns

crsp = util.comebineRet(crsp_d,dlret)

crsp.head()

### Prune Stocks

# Prune stocks, map all equity to largest permno
crsp = util.maxCap(crsp)

crsp.head()

# Trim daily returns to only those companies (permno) that have relevant market cap on dec 12 of that year
crsp = crsp[crsp['permno'].isin(permshort)]

crsp.head()

### Merge in S&P 500

# Merge crsp and sp500
crsp = crsp.merge(crsp_sp,on=['jdate'])

### Top X by Market Cap

crsp = util.topX(crsp,'me',numstocks)

# Clear out any offending zero columns
crsp = crsp.dropna(how='all',axis=1)

# Clear out any remianing zero rows
crsp = crsp.dropna(how='any',axis=0)

crsp.head()

crsp.sprtrn.describe()

### Loop Prep

# Do this stuff before the loop to save time
crsp['month'] = crsp['jdate'].dt.month
crsp['year'] = crsp['jdate'].dt.year
permlist['month'] = permlist['jdate'].dt.month
permlist['year'] = permlist['jdate'].dt.year
permlist['mktSD'] = 0
permlist['covmktstock'] = 0
permlist['beta'] = 0
crsp = crsp.sort_values(['permno','jdate'])
crsp = crsp.reset_index(drop=True)

# Details the specifics of the type of winsorized beta
delta = 3
refsig = 'sprtrn'
wsig = 'retadj'
df = util.winsorize(crsp,wsig,refsig,delta)

# Adjust returns to be used in loop
df.head()

to get a normal beta you would just not adjust the returns

#### Save Files "Pre For Loop"

crsp.to_csv("prefor_crsp.csv.gz", 
           index=False, 
           compression="gzip")

permlist.to_csv("prefor_permlist.csv.gz", 
           index=False, 
           compression="gzip")

##### can read back in "pre-loop" files if need be

## Beta Loop

# Initialize a dataframe
looplist = crsp['permno'].drop_duplicates().to_list()

# Loop to calculate betas directly using 
ll = len(looplist)

for i in range(ll):
    if(i%10==0):
        print(str(round(i/ll*100,2)) + '% complete')
        
    #Initialize loop dataframe
    loop_df = crsp[crsp['permno']==looplist[i]]
    year_loop_iter = permlist[permlist['permno']==looplist[i]]['year']

    for yiter in year_loop_iter:
        try:
            idx = permlist.loc[(permlist['permno']==looplist[i])
                &(permlist['year']==yiter)
                &(permlist['month']==12)].index[0]
        except:
            pass
        subloop_df = loop_df[(loop_df['year']>yiter-window)&(loop_df['year']<=yiter)]
        mktSD = subloop_df['sprtrn'].std()
        covmktstock = subloop_df['sprtrn'].cov(subloop_df['wretadj'])
        beta = covmktstock/mktSD**2
        permlist.loc[idx,'mktSD'] = mktSD
        permlist.loc[idx,'covmktstock'] = covmktstock
        permlist.loc[idx,'beta'] = beta

permlist.head(100)

### Export

permlist.to_csv("winbeta_"+str(window)+".csv.gz", 
           index=False, 
           compression="gzip")
