==> Code/4b - Neglected Beta Fama-MacBeth.ipynb <==
## Better Beta Fama-MacBeth

### Import Packages

import pandas as pd
import numpy as np
import datetime as dt
import psycopg2 
import matplotlib.pyplot as plt
from dateutil.relativedelta import *
from pandas.tseries.offsets import *
from scipy import stats
import statsmodels.api as sm
import statistics
import pickle
import sys
sys.path.insert(0, "../")
import util
from importlib import reload
util=reload(util)

### Read in util.py Function Bank

from IPython.lib.backgroundjobs import BackgroundJobFunc

with open('util.py') as code:
    job = BackgroundJobFunc(exec, code.read())

result = job.run()

### Import Data

#### Annual Returns

annRet = pd.read_csv('annRtnlagwBeta.csv', usecols=lambda x: x.lower())

annRet.head()

annRet.dtypes

# Drop and rename
annRet.drop(columns=['Unnamed: 0','PERMCO'],inplace=True)

annRet.rename(columns={'PERMNO': 'permno', 'YEAR': 'year', 'CompoundRet': 'annRet', 'wbeta': 'wBeta'}, inplace=True)

annRet.annRet.describe()

annRet[annRet['']]

#### Better Beta Permlist

betterbetadf = pd.read_csv('permlist_better_beta.csv.gz', compression='gzip', usecols=lambda x: x.lower())

betterbetadf.head()

betterbetadf[(betterbetadf['permno']==14596) & (betterbetadf['year']>=2019)]

betterbetadf.beta_delta.describe()

betterbetadf[betterbetadf['better_beta_flag']==1]['beta_delta'].describe()

betterbetadf[betterbetadf['better_beta_flag']==1]['beta_delta'].describe()

betterbetadf.permno.nunique()

betterbetadf.shape

# Beta stats after looking at historical outliers
betterbetadf.win_beta.describe()

# Beta stats after looking at historical outliers
betterbetadf.normal_beta.describe()

betterbetadf.head()

### Annualized Returns

AnnRet2 = pd.read_csv('AnnRet.csv.gz', compression='gzip', usecols=lambda x: x.lower())

AnnRet2['jdate'] = pd.to_datetime(AnnRet2['jdate'])

AnnRet2['annRetadj'] = AnnRet2['annRet']-0.05

AnnRet2.head()

# from these files, already have top 3k by market share every december
# permno already mapped to permco
# just need to merge, lag, split by quantile and regress

AnnRet2.permno.nunique()

AnnRet2.annRet.describe()

**Should be around 10%...**

# Clean up pre-merge
betterbetadf.drop(columns=['month','year'],inplace=True)

betterbetadf.permno.nunique()

crsp = AnnRet.merge(betterbetadf,how='inner',on=['permno','jdate'])

crsp.permno.nunique()

crsp.annRet.describe()

crsp.head()

crsp = util.lag(crsp,'beta_delta')

crsp.permno.nunique()

crsp[crsp['permno']==70519].head()

crsp.dtypes

crsp['me'] = crsp['shrout'] * crsp['prc']

crsp[['permno', 'jdate', 'shrout', 'prc', 'me']].head()

crsp.permno.nunique()

crsp = crsp.dropna(subset = ['lagbeta_delta'])

crsp.permno.nunique()

# Pick top xxxx based on market cap per jdate, which is annual now
#crsp = util.topX(crsp, 'me', 2000)

def topX(dfinput, tsig,nstocks):
    df=dfinput.copy()
    #rank all stocks based on signal 'tsig', e.g. market cap or beta etc at given time
    df[tsig+'Rank']=df.groupby(['jdate'])[tsig].rank(ascending=False)
    #drop all stocks higher rank than that signal value 
    df.drop(df.loc[df[tsig+'Rank']>nstocks].index,inplace=True)
    df=df.drop([tsig+'Rank'],axis=1)
    return df

df=crsp.copy()
#rank all stocks based on signal 'tsig', e.g. market cap or beta etc at given time
df['me_Rank']=df.groupby(['jdate', 'permno'])['me'].rank(ascending=False)
df = df.sort_values(['jdate', 'me_Rank'], ascending=[True, True])
#drop all stocks higher rank than that signal value 
#df.drop(df.loc[df['me'+'Rank']>3000],inplace=True)
#df=df.drop([tsig+'Rank'],axis=1)

df[['permno', 'jdate', 'me', 'me_Rank', 'lagbeta_delta', 'annRet']].head(50)

crsp.permno.nunique()

crsp.head()

# Review of Citi = 70519
crsp[crsp['permno']==70519]

crsp[crsp['permno']==70519].annRet.describe()

# ServiceNow Check = 13447
crsp[crsp['permno']==13447].annRet.describe()

crsp.annRet.describe()

crsp.lagbeta_delta.describe()

crsp_sorted = crsp.sort_values('lagbeta_delta', ascending=False)

crsp_sorted.head()

crsp_sorted.permno.nunique()

### Better Beta Regression w/o Quantiles

df_ols = crsp_sorted.copy()

import statsmodels.formula.api as smf

def ols_coef(x,formula):
    return smf.ols(formula,data=x).fit().params

gamma = (df_ols.groupby(['jdate']).apply(ols_coef,'annRet ~ lagbeta_delta'))

def fm_summary(p):
    s = p.describe().T
    s['std_error'] = s['std']/np.sqrt(s['count'])
    s['tstat'] = s['mean']/s['std_error']
    return s[['mean','std_error','tstat']]

fm_summary(gamma)

### Better Beta Regression w/Quantiles

df = crsp_sorted.copy()
#df = returns1.copy()

qsig = 'lagbeta_delta'
#qsig = 'lagstockSD'

nquant = 5
df = util.splitQ(df,qsig,nquant)

df.head()

temp = {}

for i in range(nquant):
    dftemp = df[df[qsig+'q']==i]
    import statsmodels.formula.api as smf

    def ols_coef(x,formula):
        return smf.ols(formula,data=x).fit().params

    gamma = (dftemp.groupby(['jdate']).apply(ols_coef,'annRet ~ lagbeta_delta'))

    def fm_summary(p):
        s = p.describe().T
        s['std_error'] = s['std']/np.sqrt(s['count'])
        s['tstat'] = s['mean']/s['std_error']
        return s[['mean','std_error','tstat']]

    temp['q='+str(i)]=fm_summary(gamma)
    
print(str(dftemp.groupby(['jdate'])['permno'].count()[0])+' stocks in each quantile')
for entry in temp:
    print('\n'+entry,'\n',temp[entry])

dftemp.columns

dftemp.head()

dftemp[dftemp['lagbeta_deltaq']==4]

dftemp.annRet.describe()

dftemp.to_csv("qFFdata_better_beta.csv.gz", 
           index=False, 
           compression="gzip")

for k in temp.keys():
    print(temp[k])

### Fama-MacBeth Gamma Plot

# Prep data and create beta quantiles

df2 = crsp.copy()
#df = returns1.copy()

qsig = 'lagbeta_delta'
#qsig = 'lagstockSD'

nquant = 5
df2 = util.splitQ(df2,qsig,nquant)

# Start by creating an empty dataframe
coef_table = pd.DataFrame()

for i in range(1991, 2021):
    print("Year: ", i)
    
    # Subset data for each year and the largest quantile
    dftemp2 = df2[(df2.year == i) & (df2.lagbeta_deltaq == 4)]
    
    y = dftemp2[['annRet']]
    X = sm.add_constant(dftemp2[['lagbeta_delta']])
    model_FM = sm.OLS(y,X).fit()
    params = pd.Series(model_FM.params)
    print(i, " ", params)
    coef_table = coef_table.append(params, ignore_index=True)

# Change column names
#coef_table.rename({0: 'Intercept', 1: 'Gamma'}, axis=1, inplace=True)
coef_table['year'] = range(1991, 2021)
print(coef_table)

coef_table.describe()

#### T-Stat of Yearly Signal

# Run T-Test on Coefficients' Time Series
print("Mean: ", coef_table[['lagbeta_delta']].mean())
print("Standard Deviation: ", coef_table[['lagbeta_delta']].std())
print("T: ", coef_table[['lagbeta_delta']].mean()/coef_table[['lagbeta_delta']].std()*(coef_table.shape[0]-1)**(1/2))

#### Plot

plt.plot(coef_table.year, coef_table.lagbeta_delta)
plt.title('Gamma Over Time')
plt.xlabel('Year')
plt.ylabel('Coefficient of Beta')
plt.show()
