==> Code/4c - Sector Definitions.ipynb <==
## Sector Definitions
### Import Packages

import pandas as pd
import numpy as np
import datetime as dt
import wrds
import psycopg2 
import matplotlib.pyplot as plt
from dateutil.relativedelta import *
from pandas.tseries.offsets import *
from scipy import stats
import statsmodels.api as sm
import statistics
import pickle
import sys
sys.path.insert(0, "../")
import util
from importlib import reload
util=reload(util)

conn = wrds.Connection(wrds_username='natejensen')

crsp_cd = conn.raw_sql("""
                      select permno, date, comnam, cusip, ticker, siccd
                      from crsp.mse
                      where exchcd between 1 and 3
                      and shrcd between 10 and 11
                      """) 

crsp_cd['date'] = pd.to_datetime(crsp_cd['date'])

crsp_cd['sicdate'] = crsp_cd['date']+MonthEnd(0)

# Initialize table of industry codes
indCodes = pd.DataFrame(columns=['siccd','indCode'])

# Create a table of industry codes
for sic in range(100,10000):
    
# 1 NoDur  Consumer NonDurables -- Food, Tobacco, Textiles, Apparel, Leather, Toys
    if (sic >=100)&(sic<=999): 
        indCode = 1
        indCodes = indCodes.append({'siccd':sic,
                                   'indCode':indCode},
                                  ignore_index=True)
    elif (sic >=2000)&(sic<=2399): 
        indCode = 1
        indCodes = indCodes.append({'siccd':sic,
                                   'indCode':indCode},
                                  ignore_index=True)
    elif (sic >=2700)&(sic<=2749): 
        indCode = 1
        indCodes = indCodes.append({'siccd':sic,
                                   'indCode':indCode},
                                  ignore_index=True)   
    elif (sic >=2770)&(sic<=2799): 
        indCode = 1
        indCodes = indCodes.append({'siccd':sic,
                                   'indCode':indCode},
                                  ignore_index=True) 
    elif (sic >=3100)&(sic<=3199): 
        indCode = 1
        indCodes = indCodes.append({'siccd':sic,
                                   'indCode':indCode},
                                  ignore_index=True) 
    elif (sic >=3940)&(sic<=3989): 
        indCode = 1
        indCodes = indCodes.append({'siccd':sic,
                                   'indCode':indCode},
                                  ignore_index=True) 
# 2 Durbl  Consumer Durables -- Cars, TVs, Furniture, Household Appliances
    elif (sic >=2500)&(sic<=2519): 
        indCode = 2
        indCodes = indCodes.append({'siccd':sic,
                                   'indCode':indCode},
                                  ignore_index=True) 
    elif (sic >=2590)&(sic<=2599): 
        indCode = 2
        indCodes = indCodes.append({'siccd':sic,
                                   'indCode':indCode},
                                  ignore_index=True)
    elif (sic >=3630)&(sic<=3659): 
        indCode = 2
        indCodes = indCodes.append({'siccd':sic,
                                   'indCode':indCode},
                                  ignore_index=True)
    elif (sic >=3710)&(sic<=3711): 
        indCode = 2
        indCodes = indCodes.append({'siccd':sic,
                                   'indCode':indCode},
                                  ignore_index=True)
    elif (sic >=3714)&(sic<=3714): 
        indCode = 2
        indCodes = indCodes.append({'siccd':sic,
                                   'indCode':indCode},
                                  ignore_index=True)
    elif (sic >=3716)&(sic<=3716): 
        indCode = 2
        indCodes = indCodes.append({'siccd':sic,
                                   'indCode':indCode},
                                  ignore_index=True)
    elif (sic >=3750)&(sic<=3751): 
        indCode = 2
        indCodes = indCodes.append({'siccd':sic,
                                   'indCode':indCode},
                                  ignore_index=True) 
    elif (sic >=3792)&(sic<=3792): 
        indCode = 2
        indCodes = indCodes.append({'siccd':sic,
                                   'indCode':indCode},
                                  ignore_index=True) 
    elif (sic >=3900)&(sic<=3939): 
        indCode = 2
        indCodes = indCodes.append({'siccd':sic,
                                   'indCode':indCode},
                                  ignore_index=True) 
    elif (sic >=3990)&(sic<=3999): 
        indCode = 2
        indCodes = indCodes.append({'siccd':sic,
                                   'indCode':indCode},
                                  ignore_index=True) 
        
# 3 Manuf  Manufacturing -- Machinery, Trucks, Planes, Off Furn, Paper, Com Printing
    elif (sic >=2520)&(sic<=2589): 
        indCode = 3
        indCodes = indCodes.append({'siccd':sic,
                                   'indCode':indCode},
                                  ignore_index=True)
    elif (sic >=2600)&(sic<=2699): 
        indCode = 3
        indCodes = indCodes.append({'siccd':sic,
                                   'indCode':indCode},
                                  ignore_index=True)
    elif (sic >=2750)&(sic<=2769): 
        indCode = 3
        indCodes = indCodes.append({'siccd':sic,
                                   'indCode':indCode},
                                  ignore_index=True)
    elif (sic >=3000)&(sic<=3099): 
        indCode = 3
        indCodes = indCodes.append({'siccd':sic,
                                   'indCode':indCode},
                                  ignore_index=True)
    elif (sic >=3200)&(sic<=3569): 
        indCode = 3
        indCodes = indCodes.append({'siccd':sic,
                                   'indCode':indCode},
                                  ignore_index=True)
    elif (sic >=3580)&(sic<=3629): 
        indCode = 3
        indCodes = indCodes.append({'siccd':sic,
                                   'indCode':indCode},
                                  ignore_index=True)
    elif (sic >=3700)&(sic<=3709): 
        indCode = 3
        indCodes = indCodes.append({'siccd':sic,
                                   'indCode':indCode},
                                  ignore_index=True)
    elif (sic >=3712)&(sic<=3713): 
        indCode = 3
        indCodes = indCodes.append({'siccd':sic,
                                   'indCode':indCode},
                                  ignore_index=True)
    elif (sic >=3715)&(sic<=3715): 
        indCode = 3
        indCodes = indCodes.append({'siccd':sic,
                                   'indCode':indCode},
                                  ignore_index=True)
    elif (sic >=3717)&(sic<=3749): 
        indCode = 3
        indCodes = indCodes.append({'siccd':sic,
                                   'indCode':indCode},
                                  ignore_index=True)
    elif (sic >=3752)&(sic<=3791): 
        indCode = 3
        indCodes = indCodes.append({'siccd':sic,
                                   'indCode':indCode},
                                  ignore_index=True)
    elif (sic >=3793)&(sic<=3799): 
        indCode = 3
        indCodes = indCodes.append({'siccd':sic,
                                   'indCode':indCode},
                                  ignore_index=True)
    elif (sic >=3830)&(sic<=3839): 
        indCode = 3
        indCodes = indCodes.append({'siccd':sic,
                                   'indCode':indCode},
                                  ignore_index=True)
    elif (sic >=3860)&(sic<=3899): 
        indCode = 3
        indCodes = indCodes.append({'siccd':sic,
                                   'indCode':indCode},
                                  ignore_index=True)
        
# 4 Enrgy  Oil, Gas, and Coal Extraction and Products
    elif (sic >=1200)&(sic<=1399): 
        indCode = 4
        indCodes = indCodes.append({'siccd':sic,
                                   'indCode':indCode},
                                  ignore_index=True)
    elif (sic >=2900)&(sic<=2999): 
        indCode = 4
        indCodes = indCodes.append({'siccd':sic,
                                   'indCode':indCode},
                                  ignore_index=True)
        
# 5 Chems  Chemicals and Allied Products
    elif (sic >=2800)&(sic<=2829): 
        indCode = 5
        indCodes = indCodes.append({'siccd':sic,
                                   'indCode':indCode},
                                  ignore_index=True)
    elif (sic >=2840)&(sic<=2899): 
        indCode = 5
        indCodes = indCodes.append({'siccd':sic,
                                   'indCode':indCode},
                                  ignore_index=True)
# 6 BusEq  Business Equipment -- Computers, Software, and Electronic Equipment
    elif (sic >=3570)&(sic<=3579): 
        indCode = 6
        indCodes = indCodes.append({'siccd':sic,
                                   'indCode':indCode},
                                  ignore_index=True)
    elif (sic >=3660)&(sic<=3692): 
        indCode = 6
        indCodes = indCodes.append({'siccd':sic,
                                   'indCode':indCode},
                                  ignore_index=True)
    elif (sic >=3694)&(sic<=3699): 
        indCode = 6
        indCodes = indCodes.append({'siccd':sic,
                                   'indCode':indCode},
                                  ignore_index=True)
    elif (sic >=3810)&(sic<=3829): 
        indCode = 6
        indCodes = indCodes.append({'siccd':sic,
                                   'indCode':indCode},
                                  ignore_index=True)
    elif (sic >=7370)&(sic<=7379): 
        indCode = 6
        indCodes = indCodes.append({'siccd':sic,
                                   'indCode':indCode},
                                  ignore_index=True)
        
# 7 Telcm  Telephone and Television Transmission;
    elif (sic >=4800)&(sic<=4899): 
        indCode = 7
        indCodes = indCodes.append({'siccd':sic,
                                   'indCode':indCode},
                                  ignore_index=True)
        
# 8 Utils  Utilities;
    elif (sic >=4900)&(sic<=4949): 
        indCode = 8
        indCodes = indCodes.append({'siccd':sic,
                                   'indCode':indCode},
                                  ignore_index=True)
        
# 9 Shops  Wholesale, Retail, and Some Services (Laundries, Repair Shops)
    elif (sic >=5000)&(sic<=5999): 
        indCode = 9
        indCodes = indCodes.append({'siccd':sic,
                                   'indCode':indCode},
                                  ignore_index=True)
    elif (sic >=7200)&(sic<=7299): 
        indCode = 9
        indCodes = indCodes.append({'siccd':sic,
                                   'indCode':indCode},
                                  ignore_index=True)
    elif (sic >=7600)&(sic<=7699): 
        indCode = 9
        indCodes = indCodes.append({'siccd':sic,
                                   'indCode':indCode},
                                  ignore_index=True)
        
# 10 Hlth   Healthcare, Medical Equipment, and Drugs
    elif (sic >=2830)&(sic<=2839): 
        indCode = 10
        indCodes = indCodes.append({'siccd':sic,
                                   'indCode':indCode},
                                  ignore_index=True)
    elif (sic >=3693)&(sic<=3693): 
        indCode = 10
        indCodes = indCodes.append({'siccd':sic,
                                   'indCode':indCode},
                                  ignore_index=True)
    elif (sic >=3840)&(sic<=3859): 
        indCode = 10
        indCodes = indCodes.append({'siccd':sic,
                                   'indCode':indCode},
                                  ignore_index=True)
    elif (sic >=8000)&(sic<=8099): 
        indCode = 10
        indCodes = indCodes.append({'siccd':sic,
                                   'indCode':indCode},
                                  ignore_index=True)

# 11 Money  Finance
    elif (sic >=6000)&(sic<=6999): 
        indCode = 11
        indCodes = indCodes.append({'siccd':sic,
                                   'indCode':indCode},
                                  ignore_index=True)
        
# 12 Other  Other -- Mines, Constr, BldMt, Trans, Hotels, Bus Serv, Entertainment
    elif (sic >=100)&(sic<=9999):
        indCode = 12
        indCodes = indCodes.append({'siccd':sic,
                                   'indCode':indCode},
                                  ignore_index=True)
    else:
        pass

industry = pd.DataFrame(columns=['indCode','industry','indDesc'])

industry = industry.append({'indCode':1,'industry':"NoDur",'indDesc':"Consumer NonDurables -- Food, Tobacco, Textiles, Apparel, Leather, Toys"},ignore_index=True)
industry = industry.append({'indCode':2,'industry':"Durbl",'indDesc':"Consumer Durables -- Cars, TVs, Furniture, Household Appliances"},ignore_index=True)
industry = industry.append({'indCode':3,'industry':"Manuf",'indDesc':"Manufacturing -- Machinery, Trucks, Planes, Off Furn, Paper, Com Printing"},ignore_index=True)
industry = industry.append({'indCode':4,'industry':"Enrgy",'indDesc':"Oil, Gas, and Coal Extraction and Products"},ignore_index=True)
industry = industry.append({'indCode':5,'industry':"Chems",'indDesc':"Chemicals and Allied Products"},ignore_index=True)
industry = industry.append({'indCode':6,'industry':"BusEq",'indDesc':"Business Equipment -- Computers, Software, and Electronic Equipment"},ignore_index=True)
industry = industry.append({'indCode':7,'industry':"Telcm",'indDesc':"Telephone and Television Transmission"},ignore_index=True)
industry = industry.append({'indCode':8,'industry':"Utils",'indDesc':"Utilities"},ignore_index=True)
industry = industry.append({'indCode':9,'industry':"Shops",'indDesc':"Wholesale, Retail, and Some Services (Laundries, Repair Shops)"},ignore_index=True)
industry = industry.append({'indCode':10,'industry':"Hlth",'indDesc':"Healthcare, Medical Equipment, and Drugs"},ignore_index=True)
industry = industry.append({'indCode':11,'industry':"Money",'indDesc':"Finance"},ignore_index=True)
industry = industry.append({'indCode':12,'industry':"Other",'indDesc':"Other -- Mines, Constr, BldMt, Trans, Hotels, Bus Serv, Entertainment"},ignore_index=True)

industry2 = pd.merge(indCodes,industry,how='inner',on=['indCode']).sort_values('siccd')

industry2['siccd'] = industry2['siccd'].astype(float)

industry3 = pd.merge(crsp_cd,industry2,how='inner',on=['siccd'])

industry3 = industry3.drop(columns=['date'])

industry3 = industry3.sort_values('sicdate')

industry3['enddate']=industry3.groupby(['permno'])['sicdate'].shift(-1)

# If enddate is missing then set to today date
industry3['enddate']=industry3['enddate'].fillna(pd.to_datetime('today'))

### Export Data

industry3.to_csv("IndustryCodes.csv.gz", 
           index=False, 
           compression="gzip")

industry3.head()

crsp_cd.to_csv("CompanyNames.csv.gz", 
           index=False, 
           compression="gzip")
